# docs/openapi-intranet-santa-rosa.yaml
openapi: 3.0.3
info:
  title: Intranet · Centro Médico Santa Rosa de Lima
  version: 1.0.0
  description: |
    API de ejemplo de la **intranet de gestión** del Centro Médico Santa Rosa de Lima
    (Proyecto Intermodular · DAW). Backend en **FastAPI** con autenticación **JWT**
    y permisos por **roles**.

    **Roles:** `personal` (todo el personal), `admin`, `direccion`, `supervision`,
    `mantenimiento`, `suministros` y `rrss`. El rol `admin` tiene acceso total.

    **Cómo probarla con el servidor de ejemplo:** ejecuta `POST /auth/google` con el email
    de un usuario de ejemplo en `id_token`, copia el `access_token`, pulsa **Authorize** y pégalo.
  license:
    name: MIT
servers:
  - url: http://127.0.0.1:8000/api/v1
    description: Servidor local de ejemplo (uvicorn)
...
security:
  - bearerAuth: []
paths:
...
  /auth/google:
    post:
      tags: [Autenticación]
      summary: Iniciar sesión con Google (OAuth)
      description: |
        Valida el token de Google y devuelve los tokens JWT de la intranet.
        Si el email no está registrado, se da de alta en estado `pendiente` y se responde 403.
        En el servidor de ejemplo (DEMO_MODE) se escribe el email en `id_token`.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/LoginGoogle' }
      responses:
        '200':
          description: Sesión iniciada
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Tokens' }
        '403':
          description: Usuario pendiente de confirmación o inactivo
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
              example: { detail: "Alta registrada: pendiente de confirmación por un administrador" }
        '422': { $ref: '#/components/responses/Validacion' }
  /auth/refresh:
    post:
      tags: [Autenticación]
      summary: Renovar el access token
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RefreshIn' }
      responses:
        '200':
          description: Nuevo access token
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AccessToken' }
        '401': { $ref: '#/components/responses/NoAutenticado' }
  /auth/me:
    get:
      tags: [Autenticación]
      summary: Datos del usuario autenticado
      responses:
        '200':
          description: Perfil del usuario
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Usuario' }
        '401': { $ref: '#/components/responses/NoAutenticado' }

  # ------------------------------------------------------------------ USUARIOS
...
  /usuarios/{usuario_id}:
    patch:
      tags: [Usuarios]
      summary: Modificar parcialmente un usuario
      description: |
        Rol requerido: `admin`. **PATCH** porque solo se envían los campos que cambian.
        Uso típico: aprobar un alta pendiente con `{"estado": "activo"}`.
      parameters:
        - $ref: '#/components/parameters/UsuarioId'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UsuarioPatch' }
            example: { estado: activo }
      responses:
        '200':
          description: Usuario actualizado
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Usuario' }
        '401': { $ref: '#/components/responses/NoAutenticado' }
        '403': { $ref: '#/components/responses/SinPermiso' }
        '404': { $ref: '#/components/responses/NoEncontrado' }
...
  /anuncios:
    get:
      tags: [Anuncios]
      summary: Ver el tablón de anuncios
      description: Anuncios generales y del servicio del usuario. Los caducados no se muestran.
      responses:
        '200':
          description: Anuncios visibles (destacados primero)
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Anuncio' }
        '401': { $ref: '#/components/responses/NoAutenticado' }
    post:
      tags: [Anuncios]
      summary: Publicar un anuncio
      description: 'Roles requeridos: `direccion` o `supervision`. Si es destacado, genera una notificación.'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AnuncioIn' }
      responses:
        '201':
          description: Anuncio creado
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Anuncio' }
        '401': { $ref: '#/components/responses/NoAutenticado' }
        '403': { $ref: '#/components/responses/SinPermiso' }
        '422': { $ref: '#/components/responses/Validacion' }

  # ------------------------------------------------------------------ GUARDIAS
...
  /ausencias:
    post:
      tags: [Ausencias]
      summary: Registrar una ausencia
      description: |
        Las ausencias `generica` nacen en estado `pendiente_justificante`; el resto, en `registrada`.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AusenciaIn' }
      responses:
        '201':
          description: Ausencia registrada
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Ausencia' }
        '401': { $ref: '#/components/responses/NoAutenticado' }
        '422': { $ref: '#/components/responses/Validacion' }
  /ausencias/{ausencia_id}/justificante:
    post:
      tags: [Ausencias]
      summary: Adjuntar justificante
      description: Solo para ausencias `generica` en estado `pendiente_justificante`.
      parameters:
        - $ref: '#/components/parameters/AusenciaId'
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [archivo]
              properties:
                archivo:
                  type: string
                  format: binary
                  description: PDF o imagen del justificante
      responses:
        '200':
          description: Justificante adjuntado (estado `justificada`)
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Ausencia' }
        '401': { $ref: '#/components/responses/NoAutenticado' }
        '403': { $ref: '#/components/responses/SinPermiso' }
        '404': { $ref: '#/components/responses/NoEncontrado' }
        '409': { $ref: '#/components/responses/Conflicto' }
  /ausencias/{ausencia_id}/pdf:
    get:
      tags: [Ausencias]
      summary: Descargar el documento PDF de la ausencia
      description: Autocompleta los datos personales desde el perfil (cuenta de Google vinculada).
      parameters:
        - $ref: '#/components/parameters/AusenciaId'
      responses:
        '200':
          description: Documento PDF
          content:
            application/pdf:
              schema: { type: string, format: binary }
        '401': { $ref: '#/components/responses/NoAutenticado' }
        '403': { $ref: '#/components/responses/SinPermiso' }
        '404': { $ref: '#/components/responses/NoEncontrado' }
        '409': { $ref: '#/components/responses/Conflicto' }
  /ausencias/{ausencia_id}/firma:
    post:
      tags: [Ausencias]
      summary: Firmar el documento
      description: |
        Firma `manuscrita` (imagen capturada en pantalla, en base64) o `autofirma`
        (firma electrónica, objetivo de mejor esfuerzo: 501 en el servidor de ejemplo).
      parameters:
        - $ref: '#/components/parameters/AusenciaId'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/FirmaIn' }
      responses:
        '200':
          description: Documento firmado (estado `firmada`)
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Ausencia' }
        '401': { $ref: '#/components/responses/NoAutenticado' }
        '403': { $ref: '#/components/responses/SinPermiso' }
        '404': { $ref: '#/components/responses/NoEncontrado' }
        '409': { $ref: '#/components/responses/Conflicto' }
        '422': { $ref: '#/components/responses/Validacion' }
        '501':
          description: AutoFirma no disponible
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }

  # ------------------------------------------------------------------ TICKETS
...
  /dashboard/kpis:
    get:
      tags: [Dirección]
      summary: Indicadores para la dirección
      description: 'Rol requerido: `direccion`.'
      responses:
        '200':
          description: Indicadores agregados
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Kpis' }
        '401': { $ref: '#/components/responses/NoAutenticado' }
        '403': { $ref: '#/components/responses/SinPermiso' }
  /auditoria:
    get:
      tags: [Dirección]
      summary: Consultar el registro de auditoría
      description: 'Rol requerido: `direccion`. Devuelve primero los registros más recientes.'
      parameters:
        - name: usuario_id
          in: query
          schema: { type: string }
        - name: limit
          in: query
          schema: { type: integer, minimum: 1, maximum: 500, default: 50 }
      responses:
        '200':
          description: Registros de auditoría
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/RegistroAuditoria' }
        '401': { $ref: '#/components/responses/NoAutenticado' }
        '403': { $ref: '#/components/responses/SinPermiso' }
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    UsuarioId:
      name: usuario_id
      in: path
      required: true
      schema: { type: string, example: u9 }
    AusenciaId:
      name: ausencia_id
      in: path
      required: true
      schema: { type: string, example: aus101 }
  responses:
    NoAutenticado:
      description: Falta el token o no es válido
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          example: { detail: No autenticado }
    SinPermiso:
      description: El usuario no tiene el rol necesario
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          example: { detail: 'Se requiere uno de estos roles: admin' }
    NoEncontrado:
      description: El recurso no existe
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          example: { detail: Recurso no encontrado }
    Conflicto:
      description: Conflicto con el estado actual (duplicado, solapamiento o transición no válida)
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          example: { detail: 'Transición no permitida: pendiente -> finalizado' }
    Validacion:
      description: Datos de entrada no válidos
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ValidationError' }
  schemas:
    Error:
      type: object
      properties:
        detail: { type: string }
    ValidationError:
      type: object
      properties:
        detail:
          oneOf:
            - type: string
            - type: array
              items:
                type: object
                properties:
                  loc: { type: array, items: { type: string } }
                  msg: { type: string }
                  type: { type: string }

    # ---------- Enumerados
    Franja:
      type: string
      enum: [manana, tarde, noche]
    EstadoUsuario:
      type: string
      enum: [activo, pendiente, inactivo]
...
    LoginGoogle:
      type: object
      required: [id_token]
      properties:
        id_token:
          type: string
          description: Token de Google. En el servidor de ejemplo, el email del usuario.
          example: laura.martinez@santarosadelima.es
    RefreshIn:
      type: object
      required: [refresh_token]
      properties:
        refresh_token: { type: string }
    Tokens:
      type: object
      properties:
        access_token: { type: string }
        refresh_token: { type: string }
        token_type: { type: string, example: bearer }
    AccessToken:
      type: object
      properties:
        access_token: { type: string }
        token_type: { type: string, example: bearer }

    # ---------- Usuarios
    Usuario:
      type: object
      properties:
        id: { type: string, example: u3 }
        nombre: { type: string, example: Ana García López }
        email: { type: string, format: email, example: ana.garcia@santarosadelima.es }
        servicio: { type: string, example: Urgencias }
        roles:
          type: array
          items: { type: string }
          example: [personal]
        estado: { $ref: '#/components/schemas/EstadoUsuario' }
        disponibilidad:
          type: array
          description: Huecos de guardia en los que está disponible
          items: { type: string }
          example: [lunes-noche, martes-noche]
    UsuarioPatch:
      type: object
      description: Solo se envían los campos que cambian
      properties:
        estado: { $ref: '#/components/schemas/EstadoUsuario' }
        servicio: { type: string }
...
    AnuncioIn:
      type: object
      required: [titulo, cuerpo]
      properties:
        titulo: { type: string, minLength: 3, example: Cambio de turno en Urgencias }
        cuerpo: { type: string, example: Desde el día 1 el turno de noche empieza a las 21:00. }
        servicio:
          type: string
          nullable: true
          description: null = anuncio general para todo el centro
          example: Urgencias
        destacado: { type: boolean, default: false }
        caduca: { type: string, format: date-time, nullable: true }
    Anuncio:
      allOf:
        - $ref: '#/components/schemas/AnuncioIn'
        - type: object
          properties:
            id: { type: string, example: a1 }
            autor_id: { type: string, example: u1 }
            publicado: { type: string, format: date-time }

    # ---------- Guardias
...
    AusenciaIn:
      type: object
      required: [tipo, fecha]
      properties:
        tipo: { type: string, enum: [generica, administrativa, actividad_externa] }
        fecha: { type: string, format: date, example: '2026-10-07' }
        dia_completo: { type: boolean, default: true }
        franjas:
          type: array
          description: Solo si no es día completo
          items: { $ref: '#/components/schemas/Franja' }
        motivo: { type: string, nullable: true, example: Consulta médica }
    Ausencia:
      allOf:
        - $ref: '#/components/schemas/AusenciaIn'
        - type: object
          properties:
            id: { type: string, example: aus101 }
            usuario_id: { type: string }
            estado: { type: string, enum: [pendiente_justificante, registrada, justificada, firmada] }
            justificante: { type: string, nullable: true }
            documento_generado: { type: boolean }
            firma: { type: string, nullable: true }
    FirmaIn:
      type: object
      required: [metodo]
      properties:
        metodo: { type: string, enum: [manuscrita, autofirma] }
        imagen_base64: { type: string, nullable: true, description: PNG de la firma capturada en pantalla }

    # ---------- Tickets
...
    Kpis:
      type: object
      properties:
        usuarios_por_estado: { type: object, additionalProperties: { type: integer } }
        guardias:
          type: object
          properties:
            asignadas: { type: integer }
            realizadas: { type: integer }
            cubiertas_por_otro_profesional: { type: integer }
            contadores_por_hueco: { type: object, additionalProperties: true }
        ausencias_por_tipo: { type: object, additionalProperties: { type: integer } }
        tickets_por_tipo_y_estado: { type: object, additionalProperties: { type: integer } }
        rrss_porcentaje_publicado: { type: number }
        reservas_totales: { type: integer }
        notificaciones_enviadas: { type: integer }
    RegistroAuditoria:
      type: object
      properties:
        fecha: { type: string, format: date-time }
        usuario_id: { type: string, nullable: true }
        accion: { type: string, example: asignar_guardias }
        detalle: { type: string }